fixed test cases and remove vendor folder

+++ kit/xor/xor.go
@@ -30,7 +30,7 @@ func (id Key) Bit(k int) int {
 
 // String returns a textual representation of the id
 func (id Key) String() string {
-	return fmt.Sprintf("%016x", id)
+	return fmt.Sprintf("%016x", uint64(id))
 }
 
 // String returns a textual representation of the id, truncated to the k MSBs.
@@ -44,7 +44,7 @@ type Point interface {
 	Key() Key
 }
 
-// Proximity equals the number of contiguous bits, starting from 
+// Proximity equals the number of contiguous bits, starting from
 // the metrically most significant one, that x and y share.
 func Proximity(x, y Point) int {
 	const nbits = 64
@@ -92,7 +92,7 @@ func (m *Metric) Dump() []Point {
 func (m *Metric) Copy() *Metric {
 	m_ := &Metric{
 		Point: m.Point,
-		n:    m.n,
+		n:     m.n,
 	}
 	if m.sub[0] != nil {
 		m_.sub[0] = m.sub[0].Copy()
